+++ Tide.hpp
@@ -47,8 +47,8 @@ private:
     irr::f32 tideHeight;
     std::vector<tidalHarmonic> tidalHarmonics;
     irr::f32 getTideGradient(uint64_t absoluteTime) const; //return der(TideHeight) (in ?? units)
-    uint64_t highTideTime(uint startSearchTime, int searchDirection=1) const; //Find previous or next high tide time
-    uint64_t lowTideTime(uint startSearchTime, int searchDirection=1) const; //Find previous or next low tide time
+    uint64_t highTideTime(uint64_t startSearchTime, int searchDirection=1) const; //Find previous or next high tide time
+    uint64_t lowTideTime(uint64_t startSearchTime, int searchDirection=1) const; //Find previous or next low tide time
 
 };
